requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-tenants==3.4.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
